% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPR4AMnew.R
\name{FPR4AMnew}
\alias{FPR4AMnew}
\title{Finding best gamma value for \code{AM}}
\usage{
FPR4AMnew(trait = trait, falseposrate = NULL, numreps = 100,
  fformula = NULL, availmemGb = 8, numgammas = 20, geno = NULL,
  pheno = NULL, map = NULL, Zmat = NULL, ncpu = detectCores(),
  ngpu = 0, seed = 101)
}
\arguments{
\item{trait}{the name of the column in the phenotype data file that contains the trait data. The name is case sensitive and must match exactly the column name in the phenotype data file.  This parameter must be specified.}

\item{numreps}{the number of replicates upon which to base the calculation of the false 
positive rate. We have found 100 replicates to be sufficient.}

\item{fformula}{the right hand side formula for the fixed effects part of the model.}

\item{availmemGb}{a numeric value. It specifies the amount of available memory (in Gigabytes). 
This should be set to the maximum practical value of available memory for the analysis.}

\item{numgammas}{the number of equidistant gamma values from 0 to 1 for which to calculate the false positive rate of the model building process.}

\item{geno}{the R  object obtained from running \code{\link{ReadMarker}}. This must be specified.}

\item{pheno}{the R  object  obtained  from running \code{\link{ReadPheno}}. This must be specified.}

\item{map}{the R object obtained from running \code{\link{ReadMap}}. If not specified, a generic map will 
be assumed.}

\item{Zmat}{the R object obtained from running \code{\link{ReadZmat}}. If not specified, an identity matrix will be assumed.}

\item{ncpu}{a integer  value for the number of CPU that are available for distributed computing.  The default is to determine the number of CPU automatically.}

\item{ngpu}{a integer value for the number of gpu available for computation.  The default
is to assume there are no gpu available.  
This option has not yet been implemented.}

\item{seed}{a integer value for the starting seed for the permutations.}
}
\description{
The gamma parameter in \code{AM} controls the conservativeness of the model 
building process. This function uses permutation to produce a table of 
gamma and coresponding false positive values. From this table, an optimal gamma can be selected for 
analysis.
}
\details{
This function calculates, empirically, the false positive rate, for different values of gamma. 
The gamma parameter is used in \code{\link{AM}} to control the conservativeness of the model 
building process. Gamma ranges in value from 0 to 1 with 
values close to 1 making \code{\link{AM}} more conservative.
A table of gamma and false positive rates is generated as the result from running this function. 
From this table,  a suitiable value of gamma can be chosen. 

To increase the precision of the estimates of false positive rate, increase \code{numreps}. 

To increase the number of grid points in which gamma is calculate, increase \code{numgammas}. 
There is no additional computational cost by increasing \code{numgammas}.
}
\examples{
  \dontrun{ 
  # Since the following code takes longer than 5 seconds to run, it has been tagged as dontrun. 
  # However, the code can be run by the user. 
  #

  #-------------------------
  #  Example  
  #------------------------

  # read the map 
  #~~~~~~~~~~~~~~
  
  # File is a plain space separated text file with the first row 
  # the column headings
  complete.name <- system.file('extdata', 'map.txt', 
                                   package='Eagle')
  map_obj <- ReadMap(filename=complete.name) 

  # read marker data
  #~~~~~~~~~~~~~~~~~~~~
  # Reading in a PLINK ped file 
  # and setting the available memory on the machine for the reading of the data to 8  gigabytes
  complete.name <- system.file('extdata', 'geno.ped', 
                                     package='Eagle')
  geno_obj <- ReadMarker(filename=complete.name,  type='PLINK', availmemGb=8) 
 
  # read phenotype data
  #~~~~~~~~~~~~~~~~~~~~~~~

  # Read in a plain text file with data on a single trait and two covariates
  # The first row of the text file contains the column names y, cov1, and cov2. 
  complete.name <- system.file('extdata', 'pheno.txt', package='Eagle')
  
  pheno_obj <- ReadPheno(filename=complete.name)
           

 # Calculate the false positive rate for AM for different gamma values. 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
  falseposrate <- FPR4AMnew(trait = 'y',
                fformula=c('cov1+cov2'),
                map = map_obj,
                pheno = pheno_obj,
                geno = geno_obj) 

}


}
\seealso{
\code{\link{AM}}
}
