% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgEMMA.R
\name{OpenUI}
\alias{OpenUI}
\title{Browser-based User Interface}
\usage{
OpenUI()
}
\description{
Opens a web browser to act as a user-friendly user interface to Eagle
}
\details{
\code{OpenUI} is an easy to use web-based user interface for Eagle. By clicking on the navigation 
tabs at the top of a page, data can be read and analysed. By using this user interface, a user can avoid having to write R code. 

Note that even though a web browser is being used as the user interface, everything remains local to the computer.
}
\examples{
\dontrun{
# opens a web browser 
OpenUI()
}

}

